repaired memory

+++ flash_attention_tensor_cores/tensor_flash_attention.cuh
@@ -4,7 +4,7 @@ __global__ void flash_attention_kernel(
 	const __half* __restrict__ Q,
 	const __half* __restrict__ K,
 	const __half* __restrict__ V,
-	__half* __restrict__ O,
+	float* __restrict__ O,
 	int B, 
     int H, 
     int L,
